requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib
bcrypt
python-jose
python-multipart
python-dotenv
smpplib